src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-contents-fallback-expected.txt

CONSOLE ERROR: line 63: Uncaught TypeError: target.createShadowRoot is not a function
